Correccion de errores para reserva

+++ backend_hotelera/commons/src/main/java/com/jesus/commons/dto/ReservaResponse.java
@@ -8,7 +8,7 @@ public record ReservaResponse(
 	Long habitacion,
 	Date fechaEntrada,
 	Date fecheSalida,
-	Integer noches,
+	Long noches,
 	Double total,
 	String estado
 ) {}
